+++ src/goryachev/common/util/CKit.java
@@ -2053,4 +2053,27 @@ public final class CKit
 	{
 		return x.toArray(new String[x.size()]);
 	}
+	
+	
+	/** creates a string containing the specified number of tabs */
+	public static String tabs(int count)
+	{
+		if(count <= 0)
+		{
+			return "";
+		}
+		return new SB(count).tab(count).toString();
+	}
+	
+	
+	/** creates a string containing the specified number of spaces */
+	public static String spaces(int count)
+	{
+		if(count <= 0)
+		{
+			return "";
+		}
+		return new SB(count).sp(count).toString();
+	}
+
 }