+++ src/goryachev/pretty/Config.java
@@ -7,4 +7,5 @@ package goryachev.pretty;
  */
 public class Config
 {
+	public static final int TAB_SIZE = 4;
 }