+++ test/goryachev/pretty/json/TestJsonParser.java
@@ -18,37 +18,37 @@ public class TestJsonParser
 	{
 		t(OB, "{", OE, "}");
 		t(AB, "[", AE, "]");
-		t(OB, "{", WH, " ", OE, "}", WH, " ", IG, "xx");
-		t(AB, "[", WH, " ", AE, "]", WH, " ", IG, "xx");
-		t(OB, "{", WH, " ", OE, "}");
-		t(AB, "[", WH, " ", AE, "]");
+		t(OB, "{", WS, " ", OE, "}", WS, " ", IG, "xx");
+		t(AB, "[", WS, " ", AE, "]", WS, " ", IG, "xx");
+		t(OB, "{", WS, " ", OE, "}");
+		t(AB, "[", WS, " ", AE, "]");
 		t(OB, "{", NB, "\"", NA, "name", NE, "\"", SP, ":", VA, "null", OE, "}");
 		t(IG, "hello");
 		t(OB, "{", NB, "\"", NA, "name", NE, "\"", SP, ":", VA, "true", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", VA, "true", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", VA, "true", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", AB, "[", AE, "]", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "\\nname", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "\\r", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "x\\r", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "x\\rx", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "\\rx", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(AB, "[", WH, " ", SB, "\"", ST, "1", SE, "\"", WH, " ", CA, ",", WH, " ", SB, "\"", ST, "1", SE, "\"", WH, " ", AE, "]", LB, "\n");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "\\u4fde", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "x\\u4fde", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "x\\u4fdex", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "\\u4fde", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "\\n\\u4fde", NE, "\"", WH, "  ", SP, ":", WH, "  ", SB, "\"", ST, "a string", SE, "\"", WH, "\n", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", AB, "[", AE, "]", CO, ",", WH, "\n ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", AB, "[", AE, "]", WH, "\n ", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", OB, "{", WH, "\n", NB, "\"", NA, "n", NE, "\"", SP, ":", VA, "true", OE, "}", CO, ",", WH, "\n ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", AB, "[", AE, "]", WH, "\n ", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", AB, "[", WH, "\n", VA, "1", WH, " ", CA, ",", WH, " ", VA, "2", WH, " ", AE, "]", WH, " ", OE, "}");
-		t(IG, "xx", WH, "  ", OB, "{", WH, " ", NB, "\"", NA, "name", NE, "\"", WH, "  ", SP, ":", WH, "  ", AB, "[", WH, "\n", VA, "1", WH, " ", CA, ",", WH, " ", VA, "2", WH, " ", AE, "]", WH, " ", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", VA, "true", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", VA, "true", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", AB, "[", AE, "]", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "\\nname", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "\\r", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "x\\r", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "x\\rx", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "\\rx", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(AB, "[", WS, " ", SB, "\"", ST, "1", SE, "\"", WS, " ", CA, ",", WS, " ", SB, "\"", ST, "1", SE, "\"", WS, " ", AE, "]", LB, "\n");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "\\u4fde", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "x\\u4fde", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "x\\u4fdex", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "\\u4fde", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "\\n\\u4fde", NE, "\"", WS, "  ", SP, ":", WS, "  ", SB, "\"", ST, "a string", SE, "\"", WS, "\n", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", AB, "[", AE, "]", CO, ",", WS, "\n ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", AB, "[", AE, "]", WS, "\n ", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", OB, "{", WS, "\n", NB, "\"", NA, "n", NE, "\"", SP, ":", VA, "true", OE, "}", CO, ",", WS, "\n ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", AB, "[", AE, "]", WS, "\n ", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", AB, "[", WS, "\n", VA, "1", WS, " ", CA, ",", WS, " ", VA, "2", WS, " ", AE, "]", WS, " ", OE, "}");
+		t(IG, "xx", WS, "  ", OB, "{", WS, " ", NB, "\"", NA, "name", NE, "\"", WS, "  ", SP, ":", WS, "  ", AB, "[", WS, "\n", VA, "1", WS, " ", CA, ",", WS, " ", VA, "2", WS, " ", AE, "]", WS, " ", OE, "}");
 		t(AB, "[", VA, "1", CA, ",", SB, "\"", ST, "s", SE, "\"", AE, "]");
 		t(AB, "[", VA, "1", CA, ",", SB, "\"", ST, "s", SE, "\"", CA, ",", VA, "3", AE, "]");
-		t(AB, "[", VA, "1", CA, ",", SB, "\"", ST, "1", SE, "\"", WH, " ", AE, "]", LB, "\n", IG, "ignore");
-		t(AB, "[", WH, " ", VA, "1", CA, ",", SB, "\"", ST, "1", SE, "\"", WH, " ", AE, "]", LB, "\n");
-		t(AB, "[", WH, " ", VA, "1", CA, ",", WH, " ", SB, "\"", ST, "1", SE, "\"", WH, " ", AE, "]", LB, "\n");
+		t(AB, "[", VA, "1", CA, ",", SB, "\"", ST, "1", SE, "\"", WS, " ", AE, "]", LB, "\n", IG, "ignore");
+		t(AB, "[", WS, " ", VA, "1", CA, ",", SB, "\"", ST, "1", SE, "\"", WS, " ", AE, "]", LB, "\n");
+		t(AB, "[", WS, " ", VA, "1", CA, ",", WS, " ", SB, "\"", ST, "1", SE, "\"", WS, " ", AE, "]", LB, "\n");
 		t(OB, "{", ER, "{");
 		t(IG, "ignore", LB, "\r\n", OB, "{", OE, "}");
 		t(OB, "{", OE, "}", LB, "\r\n", IG, "ignore", OB, "{", OE, "}", LB, "\r\n");
@@ -59,5 +59,6 @@ public class TestJsonParser
 	@Test
 	public void test()
 	{
+		t(OB, "{", NB, "\"", NA, "issue_002", NE, "\"", SP, ":", AB, "[", AB, "[", AB, "[", AE, "]", AE, "]", AE, "]", OE, "}");
 	}
 }