+++ test/goryachev/pretty/json/TestTypes.java
@@ -1,7 +1,7 @@
 // Copyright © 2017 Andy Goryachev <andy@goryachev.com>
 package goryachev.pretty.json;
-
 import goryachev.pretty.parser.Type;
+
 
 /**
  * Test Types.
@@ -14,6 +14,7 @@ public interface TestTypes
 	public static final Type CO = Type.COMMA;
 	public static final Type ER = Type.ERROR;
 	public static final Type IG = Type.IGNORE;
+	public static final Type IN = Type.INDENT;
 	public static final Type LB = Type.LINEBREAK;
 	public static final Type NA = Type.NAME;
 	public static final Type NB = Type.NAME_BEGIN;
@@ -25,7 +26,7 @@ public interface TestTypes
 	public static final Type SB = Type.STRING_BEGIN;
 	public static final Type SE = Type.STRING_END;
 	public static final Type VA = Type.VALUE;
-	public static final Type WH = Type.WHITESPACE;
+	public static final Type WS = Type.WHITESPACE;
 	public static final Type XB = Type.XML_TAG_OPEN;
 	public static final Type XC = Type.XML_COMMENT;
 	public static final Type XE = Type.XML_TAG_CLOSING;