+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 **0.00.14**
 - separate rtf and html copy menus
+- fixed #1 fails to format }}}
 
 **0.00.13**
 - binary display enhancement